Module hydra::boundary::bc_handler

Structs

BoundaryConditionHandler

The BoundaryConditionHandler struct is responsible for managing boundary conditions associated with specific mesh entities. It supports adding, retrieving, and applying boundary conditions for a given set of mesh entities and interacting with the linear system matrices and right-hand side vectors. Example usage:

Enums

BoundaryCondition

BoundaryCondition represents various types of boundary conditions that can be applied to mesh entities. These include:

Traits

BoundaryConditionApply

The BoundaryConditionApply trait defines the apply method, which is used to apply a boundary condition to a given mesh entity. It modifies the matrix and right-hand side (rhs) of the system based on the boundary condition type (Dirichlet, Neumann, Robin, etc.).

Type Aliases

BoundaryConditionFn