Trait hydra::boundary::bc_handler::BoundaryConditionApply

pub trait BoundaryConditionApply {
    // Required method
    fn apply(
        &self,
        entity: &MeshEntity,
        rhs: &mut MatMut<'_, f64>,
        matrix: &mut MatMut<'_, f64>,
        entity_to_index: &DashMap<MeshEntity, usize>,
        time: f64,
    );
}

The BoundaryConditionApply trait defines the apply method, which is used to apply a boundary condition to a given mesh entity.

Required Methods

fn apply( &self, entity: &MeshEntity, rhs: &mut MatMut<'_, f64>, matrix: &mut MatMut<'_, f64>, entity_to_index: &DashMap<MeshEntity, usize>, time: f64, )