Module hydra::time_stepping::ts

Time Stepping module

This module defines traits and structures used for time stepping in solving time-dependent problems, such as PDEs or ODEs. The TimeStepper trait defines an interface for different time stepping methods, while the TimeDependentProblem trait defines an interface for problems that can be solved using time stepping.

Structs

TimeSteppingError

Error type for time-stepping operations.

Traits

TimeDependentProblem

Trait representing a time-dependent problem, such as a system of ODEs or PDEs.

TimeStepper

Trait for time-stepping methods.