Trait hydra::time_stepping::ts::TimeDependentProblem

pub trait TimeDependentProblem {
    type State: Vector;
    type Time;

    // Required methods
    fn compute_rhs(
        &self,
        time: Self::Time,
        state: &Self::State,
        derivative: &mut Self::State,
    ) -> Result<(), TimeSteppingError>;
    fn initial_state(&self) -> Self::State;
    fn time_to_scalar(
        &self,
        time: Self::Time,
    ) -> <Self::State as Vector>::Scalar;
    fn get_matrix(&self) -> Box<dyn Matrix<Scalar = f64>>;
    fn solve_linear_system(
        &self,
        matrix: &mut dyn Matrix<Scalar = f64>,
        state: &mut Self::State,
        rhs: &Self::State,
    ) -> Result<(), TimeSteppingError>;
}

Trait representing a time-dependent problem, such as a system of ODEs or PDEs.

Types implementing this trait must define how to compute the right-hand side (RHS) of the system, provide the initial state, and define how to interact with the matrix and solver for implicit time-stepping schemes.

Associated Types:

State: The type representing the state of the system, which must implement Vector.

Time: The type representing time (typically f64 for real-valued time).

Required Associated Types

type State: Vector

type Time

Required Methods

fn compute_rhs( &self, time: Self::Time, state: &Self::State, derivative: &mut Self::State, ) -> Result<(), TimeSteppingError>

Computes the right-hand side (RHS) of the system at a given time.

Parameters

time: The current time in the simulation.

state: The current state of the system.

derivative: The output where the derivative (RHS) is stored.

Returns

Result<(), TimeSteppingError>: Ok if the computation was successful, otherwise an error.

fn initial_state(&self) -> Self::State

Returns the initial state of the system at the beginning of the simulation.

Returns

State: The initial state of the system.

fn time_to_scalar(&self, time: Self::Time) -> <Self::State as Vector>::Scalar

Converts time to the scalar type of the vector.

This allows for using time values in vector operations like axpy.

fn get_matrix(&self) -> Box<dyn Matrix<Scalar = f64>>

Returns a matrix representation of the system if applicable.

This function is used in implicit time-stepping methods to retrieve the matrix for the linear system that needs to be solved at each time step.

Returns

Option<dyn Matrix<Scalar = f64>>: Some matrix if available, or None if not applicable.

fn solve_linear_system( &self, matrix: &mut dyn Matrix<Scalar = f64>, state: &mut Self::State, rhs: &Self::State, ) -> Result<(), TimeSteppingError>

Solves the linear system A * x = b for implicit time-stepping methods.

Parameters

matrix: The matrix A in the system.

state: The state vector x that is being solved for.

rhs: The right-hand side vector b in the system.

Returns

Result<(), TimeSteppingError>: Ok if the system was solved successfully, otherwise an error.