Trait hydra::time_stepping::ts::TimeStepper

pub trait TimeStepper<P: TimeDependentProblem> {
    // Required methods
    fn step(
        &mut self,
        problem: &P,
        time: P::Time,
        dt: P::Time,
        state: &mut P::State,
    ) -> Result<(), TimeSteppingError>;
    fn adaptive_step(
        &mut self,
        problem: &P,
        time: P::Time,
        state: &mut P::State,
    ) -> Result<(), TimeSteppingError>;
    fn set_time_interval(&mut self, start_time: P::Time, end_time: P::Time);
    fn set_time_step(&mut self, dt: P::Time);
}

Trait for time-stepping methods.

This trait defines the interface for time-stepping methods, such as explicit methods like Forward Euler, or implicit methods like Backward Euler. The primary function is step(), which advances the solution by one time step.

Associated Types:

P: The type representing the time-dependent problem to be solved.

Required Methods

fn step( &mut self, problem: &P, time: P::Time, dt: P::Time, state: &mut P::State, ) -> Result<(), TimeSteppingError>

Performs a single time step.

Parameters

problem: The time-dependent problem being solved.

time: The current time in the simulation.

dt: The time step size.

state: The current state of the system, which will be updated in place.

Returns

Result<(), TimeSteppingError>: Ok if the time step was successful, otherwise an error.

fn adaptive_step( &mut self, problem: &P, time: P::Time, state: &mut P::State, ) -> Result<(), TimeSteppingError>

Performs an adaptive time step, if applicable.

Some time-stepping methods support adaptive time-stepping, where the time step size is dynamically adjusted based on error estimates or other criteria.

Parameters

problem: The time-dependent problem being solved.

time: The current time in the simulation.

state: The current state of the system, which will be updated in place.

Returns

Result<(), TimeSteppingError>: Ok if the time step was successful, otherwise an error.

fn set_time_interval(&mut self, start_time: P::Time, end_time: P::Time)

Sets the time interval for the simulation.

This function is used to specify the start and end times of the simulation.

Parameters

start_time: The start time of the simulation.

end_time: The end time of the simulation.

fn set_time_step(&mut self, dt: P::Time)

Sets the time step size for the simulation.

This function is used to specify the size of the time step in explicit or implicit time-stepping methods.

Parameters

dt: The time step size.